Create persistent reproducible package IDs
Do no longer use random-based IDs for synthesizing the full package
names, but rather create the ID as a truncated hash of the sorted
dependencies.

This allows to identify functionally identical metapackages with
different user-chosen name prefixes.

Also let all generated packages provide the same virtual pseudo-package,
allowing to enumerate all metapackages installed with the script.

Add option -l for displaying a list of all such packages.

In addition, add undocumented option -L for displaying a list of all
installed packages following the same naming rules as the new (and
previously installed) packages. This is only provided for
backwards-compatibility with packages installed by older versions of the
script which did not provide the virtual pseudo-package for the purpose
of safe identification.

Finally, fix a bug in the help output routines which made them fail when
an empty paragraph was output last.