Support saving the generated package files
Add option -s for specifying a directory where generated packages should
be saved in addition to just installing them on-the-fly.

By default, no such directory is set and thus no packages will be saved.

The -s setting will be saved and be made permanent for future
invocations of the script until its setting is changed again.