Support persistent hexadecimal name-suffixes
If no such suffix is present, one will be generated as it has been done
before.

But if a suffix is provided, that suffix will be used as is and no
attempt will be made to re-calculate the suffix.

This allows upgrading packages with new dependencies without changing
the package name (because of a different suffix).